+++ tests/integration/cases/call_mixed_type_dicts/Dart_call.dart
@@ -1,8 +1,8 @@
-class _MgrType { dynamic op({dynamic operation}) => null; }
+class _MgrType { dynamic run({dynamic operation}) => null; }
 class _AppType { final mgr = _MgrType(); }
 final app = _AppType();
 final my_data = null;
 void main() {
-    app.mgr.op(operation: <String, dynamic>{"type": "create", "pr_id": "pr_1", "draft": true});
-    app.mgr.op(operation: <String, dynamic>{"type": "create", "pr_id": "pr_2"});
+    app.mgr.run(operation: <String, dynamic>{"type": "create", "pr_id": "pr_1", "draft": true});
+    app.mgr.run(operation: <String, dynamic>{"type": "create", "pr_id": "pr_2"});
 }