+++ tests/integration/cases/call_mixed_type_dicts/JavaScript_call.js
@@ -1,3 +1,3 @@
 var app = new Proxy({}, {get: function g() { return new Proxy(function(){}, {get: g}); }});
-app.mgr.op({ operation: {"type": "create", "pr_id": "pr_1", "draft": true} });
-app.mgr.op({ operation: {"type": "create", "pr_id": "pr_2"} });
+app.mgr.run({ operation: {"type": "create", "pr_id": "pr_1", "draft": true} });
+app.mgr.run({ operation: {"type": "create", "pr_id": "pr_2"} });