+++ tests/integration/cases/call_mixed_type_dicts/Julia_call.jl
@@ -1,5 +1,5 @@
-struct MgrType; op; end
+struct MgrType; run; end
 struct AppType; mgr; end
 app = AppType(MgrType((args...; kwargs...) -> nothing))
-app.mgr.op(operation=Dict("type" => "create", "pr_id" => "pr_1", "draft" => true))
-app.mgr.op(operation=Dict("type" => "create", "pr_id" => "pr_2"))
+app.mgr.run(operation=Dict("type" => "create", "pr_id" => "pr_1", "draft" => true))
+app.mgr.run(operation=Dict("type" => "create", "pr_id" => "pr_2"))