+++ tests/integration/cases/call_mixed_type_dicts/Matlab_call.m
@@ -1,3 +1,3 @@
-app.mgr.op = @(varargin) [];
-app.mgr.op(struct('type', "create", 'pr_id', "pr_1", 'draft', true))
-app.mgr.op(struct('type', "create", 'pr_id', "pr_2"))
+app.mgr.run = @(varargin) [];
+app.mgr.run(struct('type', "create", 'pr_id', "pr_1", 'draft', true))
+app.mgr.run(struct('type', "create", 'pr_id', "pr_2"))