+++ tests/integration/cases/call_mixed_type_dicts/Swift_call.swift
@@ -1,5 +1,5 @@
-class _mgrType { @discardableResult func op(operation: Any = 0) -> Any { 0 } }
+class _mgrType { @discardableResult func run(operation: Any = 0) -> Any { 0 } }
 class _appType { var mgr = _mgrType() }
 let app = _appType()
-app.mgr.op(operation: ["type": "create", "pr_id": "pr_1", "draft": true]);
-app.mgr.op(operation: ["type": "create", "pr_id": "pr_2"]);
+app.mgr.run(operation: ["type": "create", "pr_id": "pr_1", "draft": true]);
+app.mgr.run(operation: ["type": "create", "pr_id": "pr_2"]);