+++ tests/integration/cases/call_mixed_type_dicts/VisualBasic_call.vb
@@ -1,7 +1,7 @@
 Imports System.Collections.Generic
 Module Check
     Class MgrType_1_
-        Public Function op(operation As Object) As Object
+        Public Function run(operation As Object) As Object
             Return Nothing
         End Function
     End Class
@@ -10,7 +10,7 @@ Module Check
     End Class
     Dim app As New AppType_0_()
     Sub _calls()
-        app.mgr.op(New Dictionary(Of String, Object) From {{"type", "create"}, {"pr_id", "pr_1"}, {"draft", True}})
-        app.mgr.op(New Dictionary(Of String, Object) From {{"type", "create"}, {"pr_id", "pr_2"}})
+        app.mgr.run(New Dictionary(Of String, Object) From {{"type", "create"}, {"pr_id", "pr_1"}, {"draft", True}})
+        app.mgr.run(New Dictionary(Of String, Object) From {{"type", "create"}, {"pr_id", "pr_2"}})
     End Sub
 End Module