+++ tests/integration/cases/call_ref_args_reused/Tcl_call.tcl
@@ -1,0 +1,10 @@
+proc process {args} {}
+set single_var [list \
+    4 \
+    5 \
+    6 \
+]
+set repeated_var 1
+process repeated_var 1
+process single_var 0
+process repeated_var 8